Added different combos
Combo - constructed for a fixed combination.

PermCombo - constructed for a combination, which might be permuted (six of a kind, no matter which kind, etc). Set points to 0 to signify that the points are the sum of the dice rolls.

Use - First call set_dice with the current dice. Then just read points and combination. Currently this selects the combination which is easiest to achieve (might introduce a smarter algorithm).